pyproject.toml
[project]
name = "signal-research"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "polars>=0.20.0",
    "pyyaml>=6.0",
    "pytest>=7.0.0",
    "pytest-mock>=3.0.0",
    "pytest-cov>=4.0.0",
    "sf-quant>=0.1.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "numpy>=1.24.0",
    "statsmodels>=0.14.0",
    "scipy>=1.10.0",
    "tqdm>=4.65.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["code"]
